requirements.txt
Django==4.2
django-auditlog
django-braces==1.15.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
django-handyhelpers==0.2.18
django-userextensions==0.3.0
djangorestframework==3.13.1
djangorestframework-filters==1.0.0.dev0
drf-spectacular==0.26.2
